Remove old ScorePlot components and rename ScorePlotC3
